Reorg Storage program to look more like the others
